Add CommandDto class and update project references
This commit introduces a new `CommandDto` class in `CommandDto.cs` within the `KeyboardHero` namespace. The class includes properties for `Name`, `Guid`, `Id`, and `Count`, as well as a method `GetPrettyName` for formatting the `Name` property. Additionally, the `KeyboardHero.csproj` file has been updated to include `CommandDto.cs` in the compilation list. There are also binary changes in `CommandHistory.cs` and `ShortcutHistoryWindowControl.xaml.cs`.
[release]